refactor(ai-utils): Extract buildAgentTools from handleAgentChat.
Request: Extract inline tool-building code from handleAgentChat into reusable buildAgentTools.
Motivation: Preparing for sub-agent tool building in Task 7 — buildAgentTools needs to be
callable independently for both top-level and sub-agent tool creation.

Changes:
- buildAgentTools.js: New file with endpoint tool building, MCP client creation, and MCP tool merging logic extracted from handleAgentChat
- handleAgentChat.js: Replaced inline tool-building code with single buildAgentTools call, removed now-unused imports (tool, jsonSchema, createMCPClient, Experimental_StdioMCPTransport, serializer)

Tags: ai-utils, buildAgentTools, handleAgentChat, refactor, extraction, sub-agents, mcp, tools
